Book Store
Developed a Book Store application using the MERN stack. The project integrated MongoDB for the backend database and Express.js with Node.js for server-side development. The frontend was developed using React.js, ensuring an interactive and dynamic user interface. The application allows users to browse, search, and purchase books, with features such as user authentication, book management, and order processing. The project focused on building a scalable, secure, and responsive platform.